refactored querybuilder to respective files
